+++ libisp/prim.hh
@@ -14,7 +14,7 @@ extern void mkprim(const char*, LISPT (*fname)(void), short, lisp_type);
 extern void mkprim(const char*, LISPT (*fname)(LISPT), short, lisp_type);
 extern void mkprim(const char*, LISPT (*fname)(LISPT, LISPT), short, lisp_type);
 extern void mkprim(const char*, LISPT (*fname)(LISPT, LISPT, LISPT), short, lisp_type);
-extern LISPT nth(LISPT list, long n);
+extern LISPT nth(LISPT list, int n);
 extern LISPT closobj(LISPT vars);
 extern LISPT mklambda(LISPT args, LISPT def, lisp_type);
 extern void init_prim(void);
